+++ Zomato_Backend/src/main/java/com/food/repository/UserRepository.java
@@ -1,9 +1,11 @@
 package com.food.repository;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.food.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     // Custom database operations if needed
 }
+
